Cleaned Authentication Code. Moved code from App.jsx to /pages/AuthPage.jsx for better structure.
